Se agrega autenticacion con google y correo electronico

+++ web/src/pages/OwnerQuestionsPage.js
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react'
 import { connect } from 'react-redux'
+import Swal from 'sweetalert2'
 
 import { fetchOwnerQuestions, deleteQuestion } from '../actions/questionActions'
 import { Question } from '../components/Question'
@@ -15,8 +16,25 @@ const OwnerQuestionsPage = ({ dispatch, loading, questions, hasErrors, redirect,
         }
     }, [redirect, dispatch, userId]);
 
-    const onDelete = (id) => {
-        dispatch(deleteQuestion(id))
+    const onDelete = (id) => {       
+        Swal.fire({
+            title: 'Delete Question?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#07711D',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes Delete'
+          }).then((result) => {
+            if(result.isConfirmed) {
+            dispatch(deleteQuestion(id))
+              Swal.fire(
+                'Deleted!',
+                'Your question has been deleted.',
+                'success'
+              )
+            }
+          })
     }
